requirements.txt
runpod>=1.6.0
torch>=2.0.0
torchvision>=0.15.0
git+https://github.com/huggingface/diffusers
transformers>=4.30.0
accelerate>=0.20.0
Pillow>=9.0.0
numpy>=1.21.0
safetensors>=0.3.0
huggingface-hub>=0.15.0
xformers>=0.0.20
sentencepiece>=0.1.99
protobuf>=3.20.0